tests/wpt/web-platform-tests/web-animations/interfaces/AnimationEffectTiming/fill.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.fill</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animationeffecttiming-fill">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const anim = createDiv(t).animate(null);
  assert_equals(anim.effect.timing.fill, 'auto');
}, 'Has the default value \'auto\'');

for (const fill of ['none', 'forwards', 'backwards', 'both']) {
  test(t => {
    const div = createDiv(t);
    const anim = div.animate({ opacity: [ 0, 1 ] }, 100);
    anim.effect.timing.fill = fill;
    assert_equals(anim.effect.timing.fill, fill, 'set fill ' + fill);
    assert_equals(anim.effect.getComputedTiming().fill, fill,
                  'getComputedTiming() after set fill ' + fill);
  }, `Can set fill to ${fill}`);
}

</script>
</body>
